requirements.txt
fastapi[all]
uvicorn
streamlit
pydantic-settings
sqlmodel
sqlalchemy
numpy
torch
transformers
pandas
Pillow
matplotlib
alembic
psycopg2-binary
bcrypt
pydantic
python-dotenv
asyncpg
passlib
opencv-python
requests
absl-py
torchvision
uuid
aiofiles